test: fold surviving Observation-presence tests into fhir_summary_generator_edge, drop the wearables test husk
fhir_summary_generator_wearables.test.ts's two remaining tests only exercised
getRemainingResourcesFromCompositionBundle's generic "is any Observation-typed
resource present" check - the wearable vendor tag on its fixture was incidental
and never inspected by that code path. Keeping the file, its describe title, and
the wearable-flavored fixture around implied wearable-specific behavior that no
longer exists there. Moves the two tests (with a simplified, vendor-tag-free
Observation fixture) into fhir_summary_generator_edge.test.ts, where
resource-presence edge cases already live, and deletes the now-empty-of-purpose
file.
