+++ tests/generators/fhir_summary_generator_edge.test.ts
@@ -1,6 +1,8 @@
 import { ComprehensiveIPSCompositionBuilder } from '../../src/generators/fhir_summary_generator';
 import { TPatient } from '../../src/types/resources/Patient';
+import { TObservation } from '../../src/types/resources/Observation';
 import { TComposition } from '../../src/types/resources/Composition';
+import { TBundle } from '../../src/types/resources/Bundle';
 import { IPSSections } from '../../src/structures/ips_sections';
 
 describe('ComprehensiveIPSCompositionBuilder edge cases', () => {
@@ -76,6 +78,49 @@ describe('ComprehensiveIPSCompositionBuilder edge cases', () => {
 
     const problemSection = composition.section?.find(s => s.code?.coding?.[0]?.code === '11450-4');
     expect(problemSection?.code?.coding?.[0]?.system).toBe('http://loinc.org');
+  });
+
+  it('does not report Observation as a missing resource type when an Observation is present', () => {
+    const builder = new ComprehensiveIPSCompositionBuilder();
+    const patient: TPatient = {
+      resourceType: 'Patient',
+      id: 'p1',
+      name: [{ family: 'Doe', given: ['John'] }],
+    };
+    const observation: TObservation = {
+      resourceType: 'Observation',
+      id: 'obs-1',
+      status: 'final',
+      subject: { reference: 'Patient/p1' },
+      code: { coding: [{ system: 'http://loinc.org', code: '8867-4', display: 'Heart rate' }] },
+      effectiveDateTime: '2026-01-03T08:00:00Z',
+      valueQuantity: { value: 75, unit: 'bpm' },
+    };
+    const bundle: TBundle = {
+      resourceType: 'Bundle',
+      type: 'collection',
+      entry: [{ resource: patient }, { resource: observation }],
+    };
+
+    const remaining = builder.getRemainingResourcesFromCompositionBundle(bundle);
+    expect(remaining).not.toContain('Observation');
+  });
+
+  it('reports Observation as missing when no Observation of any kind is present', () => {
+    const builder = new ComprehensiveIPSCompositionBuilder();
+    const patient: TPatient = {
+      resourceType: 'Patient',
+      id: 'p1',
+      name: [{ family: 'Doe', given: ['John'] }],
+    };
+    const bundle: TBundle = {
+      resourceType: 'Bundle',
+      type: 'collection',
+      entry: [{ resource: patient }],
+    };
+
+    const remaining = builder.getRemainingResourcesFromCompositionBundle(bundle);
+    expect(remaining).toContain('Observation');
   });
 });
